requirements.txt
elasticsearch
elasticsearch-dsl
pydantic
redis
eventlet
python-socketio
websocket-client
requests # for sio client
